+++ docs/src/app/features/theme/color-scheme-toggle.ts
