+++ docs/src/app/features/theme/theme-switcher.ts
