+++ docs/src/main.server.ts
@@ -1,5 +1,5 @@
 import { BootstrapContext, bootstrapApplication } from '@angular/platform-browser';
-import { App } from './app/app';
+import { App } from './app/shell/app';
 import { config } from './app/app.config.server';
 
 const bootstrap = (context: BootstrapContext) =>
