+++ docs/src/main.ts
@@ -1,6 +1,6 @@
 import { bootstrapApplication } from '@angular/platform-browser';
 import { appConfig } from './app/app.config';
-import { App } from './app/app';
+import { App } from './app/shell/app';
 
 bootstrapApplication(App, appConfig)
   .catch((err) => console.error(err));
